Move the githubSupport to a place it will be exported
