+++ src/index.ts
@@ -91,9 +91,9 @@ export * from "./common/delivery/goals/Goal";
 export * from "./common/delivery/goals/Goals";
 export * from "./common/delivery/goals/SdmGoalImplementationMapper";
 export * from "./common/delivery/goals/gitHubContext";
+export * from "./common/delivery/goals/gitHubStatusSetters";
 export * from "./common/delivery/goals/goalPreconditions";
 export * from "./common/delivery/goals/storeGoals";
-export * from "./common/delivery/goals/summarizeGoalsInGitHubStatus";
 export * from "./common/delivery/goals/common/GenericGoal";
 export * from "./common/delivery/goals/common/MessageGoal";
 export * from "./common/delivery/goals/common/commonGoals";
@@ -102,6 +102,7 @@ export * from "./common/delivery/goals/common/libraryGoals";
 export * from "./common/delivery/goals/common/npmGoals";
 export * from "./common/delivery/goals/graph/graphGoalsToSlack";
 export * from "./common/delivery/goals/support/fetchGoalsOnCommit";
+export * from "./common/delivery/goals/support/githubStatusSummarySupport";
 export * from "./common/delivery/goals/support/logInterpreters";
 export * from "./common/delivery/goals/support/reportGoalError";
 export * from "./common/listener/ArtifactListener";
@@ -110,6 +111,7 @@ export * from "./common/listener/ChannelLinkListenerInvocation";
 export * from "./common/listener/ClosedIssueListener";
 export * from "./common/listener/DeploymentListener";
 export * from "./common/listener/FingerprintDifferenceListener";
+export * from "./common/listener/FingerprintListener";
 export * from "./common/listener/GoalSetter";
 export * from "./common/listener/GoalsSetListener";
 export * from "./common/listener/IssueListenerInvocation";
